requirements.txt
aiofiles==22.1.0
aiosqlite==0.18.0
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==22.2.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
blis==0.7.9
Bottleneck==1.3.7
Brotli==1.0.9
CacheControl==0.12.11
cachetools==5.3.0
catalogue==2.0.8
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
comm==0.1.2
confection==0.0.4
contourpy==1.0.7
cryptography==39.0.2
cycler==0.11.0
cymem==2.0.7
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
executing==1.2.0
fastai==1.0.60
fastjsonschema==2.16.3
fastprogress==1.0.3
ffmpeg==1.4
ffmpeg-python==0.2.0
firebase-admin==6.1.0
Flask==2.2.3
Flask-Reuploaded==1.3.0
Flask-WTF==1.1.1
fonttools==4.39.0
fqdn==1.5.1
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.80.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.2
google-cloud-firestore==2.10.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.3
grpcio-status==1.51.3
httplib2==0.21.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
ipykernel==6.21.3
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-ydoc==0.2.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.4.0
jupyter_server_fileid==0.8.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.6.1
jupyterlab==3.6.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
jupyterlab_server==2.20.0
kiwisolver==1.4.4
langcodes==3.3.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
msgpack==1.0.4
murmurhash==1.0.9
mutagen==1.46.0
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.3
notebook_shim==0.2.2
numexpr==2.8.4
numpy==1.21.5
nvidia-ml-py3==7.352.0
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathlib==1.0.1
pathtools==0.1.2
pathy==0.10.1
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.1.1
preshed==3.0.8
prometheus-client==0.16.0
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodomex==3.17
pydantic==1.10.5
Pygments==2.14.0
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2022.7.1
pywin32==305
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.0
replicate==0.5.0
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scipy==1.10.1
Send2Trash==1.8.0
sentry-sdk==1.16.0
setproctitle==1.3.2
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
stack-data==0.6.2
tensorboardX==2.6
terminado==0.17.1
thinc==8.1.9
tinycss2==1.2.1
tomli==2.0.1
torch==1.11.0+cu113
torchvision==0.12.0+cu113
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typer==0.7.0
typing_extensions==4.5.0
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.14
wandb==0.13.11
wasabi==1.1.1
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
websockets==10.4
Werkzeug==2.2.3
widgetsnbextension==4.0.5
wincertstore==0.2
WTForms==3.0.1
y-py==0.5.9
ypy-websocket==0.8.2
yt-dlp==2023.3.4
zipp==3.15.0
